[SYS_FS] Regenerated demo for fs__sdspi_fat_spi_drv__firmware__sam_d21_xpro.X
MH3-103728

+++ apps/fs/sdspi_fat_spi_drv/firmware/src/config/sam_d21_xpro/driver/sdspi/src/drv_sdspi.c
@@ -2563,7 +2563,7 @@ static void DRV_SDSPI_SetupXfer(
         {
             return;
         }
-        if (((blockStart + nBlocks) > dObj->mediaGeometryTable[SYS_MEDIA_GEOMETRY_TABLE_READ_ENTRY].numBlocks))
+        if ((((uint64_t)blockStart + nBlocks) > dObj->mediaGeometryTable[SYS_MEDIA_GEOMETRY_TABLE_READ_ENTRY].numBlocks))
         {
             return;
         }
@@ -2574,7 +2574,7 @@ static void DRV_SDSPI_SetupXfer(
         {
             return;
         }
-        if (((blockStart + nBlocks) > dObj->mediaGeometryTable[SYS_MEDIA_GEOMETRY_TABLE_WRITE_ENTRY].numBlocks))
+        if ((((uint64_t)blockStart + nBlocks) > dObj->mediaGeometryTable[SYS_MEDIA_GEOMETRY_TABLE_WRITE_ENTRY].numBlocks))
         {
             return;
         }
